refactor: update the folder structures
